+++ packages/color-helpers/src/conversions/lch-to-lab.ts
@@ -1,0 +1,13 @@
+/**
+ * @license W3C https://www.w3.org/Consortium/Legal/2015/copyright-software-and-document
+ *
+ * @copyright This software or document includes material copied from or derived from https://github.com/w3c/csswg-drafts/blob/main/css-color-4/conversions.js. Copyright © 2022 W3C® (MIT, ERCIM, Keio, Beihang).
+ */
+export function LCH_to_Lab(LCH: Color): Color {
+	// Convert from polar form
+	return [
+		LCH[0], // L is still L
+		LCH[1] * Math.cos(LCH[2] * Math.PI / 180), // a
+		LCH[1] * Math.sin(LCH[2] * Math.PI / 180), // b
+	];
+}